Adjusted comments
