+++ python/pyspark/sql/tests/connect/test_connect_creation.py
@@ -554,79 +554,6 @@ class SparkConnectCreationTests(SparkConnectSQLTestCase):
                 self.spark.createDataFrame(pdf).collect(),
             )
 
-    def test_schema_inference_from_pandas_with_dict(self):
-        # SPARK-47543: test for verifying if inferring `dict` as `MapType` work properly.
-        pdf = pd.DataFrame({"str_col": ["second"], "dict_col": [{"first": 0.7, "second": 0.3}]})
-
-        infer_pandas_dict_as_map = self.connect.conf.get(
-            "spark.sql.execution.pandas.inferPandasDictAsMap"
-        )
-        self.connect.conf.set("spark.sql.execution.pandas.inferPandasDictAsMap", True)
-
-        with self.sql_conf(
-            {
-                "spark.sql.execution.arrow.pyspark.enabled": True,
-                "spark.sql.execution.arrow.pyspark.fallback.enabled": False,
-                "spark.sql.execution.pandas.inferPandasDictAsMap": True,
-            }
-        ):
-            sdf = self.spark.createDataFrame(pdf)
-            cdf = self.connect.createDataFrame(pdf)
-            self.assertEqual(
-                sdf.withColumn("test", SF.col("dict_col")[SF.col("str_col")]).collect(),
-                cdf.withColumn("test", CF.col("dict_col")[CF.col("str_col")]).collect(),
-            )
-
-            # Empty dict should fail
-            pdf_empty_struct = pd.DataFrame({"str_col": ["second"], "dict_col": [{}]})
-
-            with self.assertRaises(PySparkValueError) as pe:
-                self.connect.createDataFrame(pdf_empty_struct)
-
-            self.check_error(
-                exception=pe.exception,
-                error_class="CANNOT_INFER_EMPTY_SCHEMA",
-                message_parameters={},
-            )
-
-            with self.assertRaises(PySparkValueError) as pe:
-                self.spark.createDataFrame(pdf_empty_struct)
-
-            self.check_error(
-                exception=pe.exception,
-                error_class="CANNOT_INFER_EMPTY_SCHEMA",
-                message_parameters={},
-            )
-
-            # Dict has different types of values should fail
-            pdf_different_type = pd.DataFrame(
-                {"str_col": ["second"], "dict_col": [{"first": 0.7, "second": "0.3"}]}
-            )
-            self.assertRaises(
-                PySparkValueError, lambda: self.connect.createDataFrame(pdf_different_type)
-            )
-            self.assertRaises(
-                PySparkValueError, lambda: self.spark.createDataFrame(pdf_different_type)
-            )
-
-        with self.sql_conf(
-            {
-                "spark.sql.execution.arrow.pyspark.enabled": False,
-                "spark.sql.execution.pandas.inferPandasDictAsMap": True,
-            }
-        ):
-            sdf = self.spark.createDataFrame(pdf)
-            cdf = self.connect.createDataFrame(pdf)
-            self.assertEqual(
-                sdf.withColumn("test", SF.col("dict_col")[SF.col("str_col")]).collect(),
-                cdf.withColumn("test", CF.col("dict_col")[CF.col("str_col")]).collect(),
-            )
-
-        # Reset config
-        self.connect.conf.set(
-            "spark.sql.execution.pandas.inferPandasDictAsMap", infer_pandas_dict_as_map
-        )
-
     def test_schema_has_nullable(self):
         schema_false = StructType().add("id", IntegerType(), False)
         cdf1 = self.connect.createDataFrame([[1]], schema=schema_false)
